Use semi-analytic solution to accelerate model spinup

This functions approximates landscape equilibrium steady state for vegetation and soil pools using the successional trajectory of a single patch modeled with disturbance off and the prescribed disturbance rates for runs (Xia et al. 2012 GMD 5:1259-1271).

SAS.ED2(
  dir.analy,
  dir.histo,
  outdir,
  lat,
  lon,
  blckyr,
  prefix,
  treefall,
  param.args = SAS.ED2.param.Args(),
  sufx = "g01.h5"
)

Arguments

dir.analy

Location of ED2 analysis files; expects monthly and yearly output

dir.histo

Location of ED2 history files (for vars not in analy); expects monthly

outdir

Location to write SAS .css & .pss files

lat

site latitude; used for file naming

lon

site longitude; used for file naming

blckyr

Number of years between patch ages (aka blocks)

prefix

ED2 -E- output file prefix

treefall

Value to be used for TREEFALL_DISTURBANCE_RATE in ED2IN for full runs (disturbance on)

param.args

ED2 parameter arguments (mostly soil biogeochem)

sufx

ED2 out file suffix; used in constructing file names(default "g01.h5)

Author

Christine Rollinson, modified from original by Jaclyn Hatala-Matthes (2/18/14) 2014 Feb: Original ED SAS solution Script at PalEON modeling HIPS sites (Matthes) 2015 Aug: Modifications for greater site flexibility & updated ED 2016 Jan: Adaptation for regional-scale runs (single-cells run independently, but executed in batches) 2018 Jul: Conversion to function, Christine Rollinson July 2018