met2model wrapper for ED2

If files already exist in 'Outfolder', the default function is NOT to overwrite them and only gives user the notice that file already exists. If user wants to overwrite the existing files, just change overwrite statement below to TRUE.

Usage

met2model.ED2(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  lst = 0,
  lat = NA,
  lon = NA,
  overwrite = FALSE,
  verbose = FALSE,
  leap_year = TRUE,
  ...
)

Arguments

in.path

location on disk where inputs are stored

in.prefix

prefix of input and output files

outfolder

location on disk where outputs will be stored

start_date

the start date of the data to be downloaded (will only use the year part of the date)

end_date

the end date of the data to be downloaded (will only use the year part of the date)

lst

timezone offset to GMT in hours

lat

latitude; if not provide the function will attempt to discover it in input files

lon

longitude; if not provide the function will attempt to discover it in input files

overwrite

should existing files be overwritten

verbose

should the function be very verbose

leap_year

Enforce Leap-years? If set to TRUE, will require leap years to have 366 days. If set to false, will require all years to have 365 days. Default = TRUE.

...

currently unused