Function for reading -T- files

Function for reading -T- files

Usage

read_T_files(
  yr,
  yfiles,
  h5_files,
  outdir,
  start_date,
  end_date,
  pfts = NULL,
  settings = NULL
)

Arguments

yr

the year being processed

yfiles

the years on the filenames, will be used to matched h5_files for that year

h5_files

names of T files to be read

outdir

directory where ED2 output files are found

start_date

start date in YYYY-MM-DD format

end_date

end date in YYYY-MM-DD format

pfts

for consistency with read_E_files()—unused

settings

A PEcAn settings object. Values for start_date and end_date will be taken from settings if it is supplied.

Details

e.g. yr = 1999 yfiles = 1999 2000 h5_files = "analysis-T-1999-00-00-000000-g01.h5" "analysis-T-2000-00-00-000000-g01.h5"