Read ED2 vegetation inputs

Read ED2 css, pss, and site files into a single ED input object.

read_ed_veg(path_prefix, latitude = NULL, longitude = NULL,
  check = TRUE)

Arguments

path_prefix
Full path and prefix to initial condition files.

latitude
Run latitude (default = NULL). If NULL, deduced from file name.

longitude
Run longitude (default = NULL). If NULL, deduced from file name.

check
Whether or not to check css, pss, and site files for validity. Default = TRUE.

Value

List containing css, pss, and site objects, latitude and longitude, and orig_paths, a list of paths to the original css, pss, and site files.